The changes I did on a plane
These changes were made on a plane with out the internet, so who knows
if they work.

1) implemented a new sessionData struct, which manages itself along with
the cookie session to store complex data structures. Will clean itself
up if not accessed every hour, or how ever long the config sets itself
for.
2) strava.js now uses the sessionData struct to store it's data
3) errorReport is a little better with reporting erros
4) stencil now doesn't crash while doing it's eval. Will log the problem
as a warning and return 'undefined'. Causing data-if's to fail, and
data-set's to be nothing.
5) Added blank gxp2segs file.
6) Error report also print's client errors, which are displayed only
once to the user.
7) Each page get's the client error box added to it if any errors
happened during the last get or use commands.